backend/requirements.txt
pdfplumber==0.11.7
fastapi==0.116.1
uvicorn[standard]==0.37.0
python-dotenv==1.1.1
requests==2.32.5
google-generativeai==0.8.5
python-multipart==0.0.20
pydantic==2.12
websockets==15.0.1
faiss-cpu==1.8.0
numpy==1.24.3
packaging==23.2
